requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiomysql==0.2.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
audioread==3.0.1
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
blinker==1.4
blis==0.7.11
catalogue==2.0.10
certifi==2024.6.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
confection==0.1.5
configobj==5.0.6
constantly==15.1.0
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
cymem==2.0.8
dbus-python==1.2.18
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
fonttools==4.53.0
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
more-itertools==8.10.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
mysql-connector==2.2.9
mysql-connector-python==8.3.0
netifaces==0.11.0
noisereduce==3.0.2
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.0
openai==1.35.4
orjson==3.10.5
packaging==24.1
pandas==2.2.2
passlib==1.7.4
pexpect==4.8.0
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.2
preshed==3.0.9
ptyprocess==0.7.0
pyasn1-modules==0.2.1
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydub==0.25.1
Pygments==2.18.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.8.0
PyMySQL==1.1.1
pyOpenSSL==21.0.0
pyparsing==3.1.2
pyrsistent==0.18.1
pyserial==3.5
python-dotenv==1.0.1
python-http-client==3.3.7
python-jose==3.3.0
python-magic==0.4.24
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
rsa==4.9
schedule==1.2.2
scikit-learn==1.5.0
scipy==1.14.0
SecretStorage==3.3.1
sendgrid==6.11.0
service-identity==18.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.31
srsly==2.4.8
ssh-import-id==5.11
starkbank-ecdsa==2.2.0
starlette==0.37.2
stripe==10.0.0
thinc==8.2.5
threadpoolctl==3.5.0
tqdm==4.66.4
twilio==9.0.4
Twisted==22.1.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
wadllib==1.3.6
watchfiles==0.22.0
weasel==0.4.1
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==1.0.0
zope.interface==5.4.0

xkit==0.0.0
uvloop==0.19.0
unattended-upgrades==0.1
ufw==0.36.1
cloud-init==24.2
command-not-found==0.3
distro-info==1.1+ubuntu0.2
python-apt==2.4.0+ubuntu4
sos==4.5.6
systemd-python==234
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
pyasn1==0.6.0
colorama==0.4.4
wasabi==1.1.3